% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hex_to_hsl}
\alias{hex_to_hsl}
\title{Convert HEX to HSL}
\usage{
hex_to_hsl(hex, ...)
}
\arguments{
\item{hex}{Character vector of hex colour strings.}

\item{...}{Additional options passed to \code{\link[=convert_colourspace]{convert_colourspace()}}.}
}
\value{
Numeric vector (length 3) or matrix with columns \code{h}, \code{s}, \code{l}.
}
\description{
Convert HEX to HSL
}
\examples{
hex_to_hsl("#336699")
}
