% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hex_to_oklab}
\alias{hex_to_oklab}
\title{Convert HEX to OKLAB}
\usage{
hex_to_oklab(hex, ...)
}
\arguments{
\item{hex}{Character vector of hex colour strings.}

\item{...}{Additional options passed to \code{\link[=convert_colourspace]{convert_colourspace()}}.}
}
\value{
Numeric vector (length 3) or matrix with columns \code{l}, \code{a}, \code{b}.
}
\description{
Convert HEX to OKLAB
}
\examples{
hex_to_oklab("#ff0000")
}
