% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hex_to_oklch}
\alias{hex_to_oklch}
\title{Convert HEX to OKLCH}
\usage{
hex_to_oklch(hex)
}
\arguments{
\item{hex}{Character vector of hex colour strings.}
}
\value{
Numeric vector (length 3) or matrix with columns \code{l}, \code{c}, \code{h}.
}
\description{
Convert HEX to OKLCH
}
\examples{
hex_to_oklch("#ff0000")
}
