% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hex_to_rgb}
\alias{hex_to_rgb}
\title{Convert HEX to RGB}
\usage{
hex_to_rgb(hex)
}
\arguments{
\item{hex}{Character vector of hex colour strings.}
}
\value{
Numeric vector (length 3) or matrix with columns \code{r}, \code{g}, \code{b}.
}
\description{
Convert HEX to RGB
}
\examples{
hex_to_rgb("#336699")
}
