% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hsl_to_hex}
\alias{hsl_to_hex}
\title{Convert HSL to HEX}
\usage{
hsl_to_hex(hsl)
}
\arguments{
\item{hsl}{Numeric vector/matrix of HSL values (h: 0-360, s/l: 0-100).}
}
\value{
Character vector of hex colours.
}
\description{
Convert HSL to HEX
}
\examples{
hsl_to_hex(c(210, 50, 40))
}
