% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{oklab_to_hex}
\alias{oklab_to_hex}
\title{Convert OKLAB to HEX}
\usage{
oklab_to_hex(oklab, ...)
}
\arguments{
\item{oklab}{Numeric vector/matrix of OKLAB values (\code{l} in 0-1).}

\item{...}{Additional options passed to \code{\link[=convert_colourspace]{convert_colourspace()}}.}
}
\value{
Character vector of hex colours.
}
\description{
Convert OKLAB to HEX
}
\examples{
oklab_to_hex(c(0.628, 0.225, 0.126))
}
