% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{oklch_to_hex}
\alias{oklch_to_hex}
\title{Convert OKLCH to HEX}
\usage{
oklch_to_hex(oklch, ...)
}
\arguments{
\item{oklch}{Numeric vector/matrix of OKLCH values (\code{l} in 0-1, \code{c} >= 0, \code{h} in degrees).}

\item{...}{Additional options passed to \code{\link[=convert_colourspace]{convert_colourspace()}}.}
}
\value{
Character vector of hex colours.
}
\description{
Convert OKLCH to HEX
}
\examples{
oklch_to_hex(c(0.628, 0.258, 29.221))
}
