% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rgb_to_hex}
\alias{rgb_to_hex}
\title{Convert RGB to HEX}
\usage{
rgb_to_hex(rgb, ...)
}
\arguments{
\item{rgb}{Numeric vector/matrix of RGB values (0-255).}

\item{...}{Additional options passed to \code{\link[=convert]{convert()}}.}
}
\value{
Character vector of hex colours.
}
\description{
Convert RGB to HEX
}
\examples{
rgb_to_hex(c(51, 102, 153))
}
