% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_names}
\alias{color_names}
\title{CSS color names to HEX mapping}
\format{
A data frame with two columns:
\describe{
\item{name}{Color name as provided by the source (character).}
\item{hex}{Lowercase hex triplet starting with '#'.}
}
}
\source{
\url{https://github.com/meodai/color-names}
}
\usage{
color_names
}
\description{
A dataset containing 31k+ color names compiled by the
\href{https://github.com/meodai/color-names}{meodai/color-names} project.
}
\keyword{datasets}
